+++ packages/twenty-front/src/modules/activities/calendar/components/CalendarEventParticipantsResponseStatusField.tsx
@@ -4,8 +4,8 @@ import styled from '@emotion/styled';
 import { IconCheck, IconQuestionMark, IconX } from 'twenty-ui';
 import { v4 } from 'uuid';
 
-import { ExpandableList } from '@/activities/calendar/components/ExpandableList/ExpandableList';
 import { CalendarEventParticipant } from '@/activities/calendar/types/CalendarEventParticipant';
+import { ExpandableList } from '@/activities/components/ExpandableList/ExpandableList';
 import { ParticipantChip } from '@/activities/components/ParticipantChip';
 import { PropertyBox } from '@/object-record/record-inline-cell/property-box/components/PropertyBox';
 import { EllipsisDisplay } from '@/ui/field/display/components/EllipsisDisplay';
