+++ packages/twenty-front/src/modules/activities/components/ExpandableList/IntersectionObserverWrapper.tsx
